+++ src/components/status_content/status_content.vue
@@ -55,12 +55,20 @@
       >
         {{ $t("status.show_content") }}
         <span
-          v-if="hasImageAttachments"
+          v-if="attachmentTypes.includes('image')"
           class="icon-picture"
         />
         <span
-          v-if="hasVideoAttachments"
+          v-if="attachmentTypes.includes('video')"
           class="icon-video"
+        />
+        <span
+          v-if="attachmentTypes.includes('audio')"
+          class="icon-music"
+        />
+        <span
+          v-if="attachmentTypes.includes('unknown')"
+          class="icon-doc"
         />
         <span
           v-if="status.card"
@@ -217,6 +225,9 @@ $status-margin: 0.75em;
     font-family: var(--postFont, sans-serif);
     line-height: 1.4em;
     white-space: pre-wrap;
+    overflow-wrap: break-word;
+    word-wrap: break-word;
+    word-break: break-word;
 
     blockquote {
       margin: 0.2em 0 0.2em 2em;